Copy the Shorted Link Using JS
copying the URL from the input field, using JavaScript.
The shorted link will be give the same result as normal link for that we use one function in that function the I get  shorted URL object and stored it on one variable then using this variable we get main and then redirect it.